singapore-tax-gpt/pyproject.toml
[project]
name = "singapore-tax-gpt"
version = "0.1.0"
description = "AI-powered Singapore tax assistant using RAG"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "chromadb>=0.4.22",
    "openai>=1.10.0",
    "streamlit>=1.29.0",
    "pypdf>=3.17.4",
    "pdfplumber>=0.10.3",
    "python-dotenv>=1.0.0",
]
